Installing texlive-latex-extra package required for PDF generation
When trying to generate a PDF in RStudio I got the following error:

    LaTeX Error: File `titling.sty' not found

Thanks to apt-file (See README) I managed to pinpoint the exact package
providing it.

NOTE: There is a TeX package which install pretty much everything but it's
big (~2GB?) and I would avoid to install it unless necessary.